feat: move cursor when input is streaming in `chat`
Previously, the cursor moved along with the messages that were streaming into the buffer. With this update, the user can navigate the buffer freely allowing for common tasks such as yanking code blocks. Should the user wish to move their cursor back to the latest message stream, they can use `G` to go to the last line in the buffer