feat: :sparkles: improved prompts
This is more significant an improvement than initially meets the eye.
I've spent the last few days curating the default system prompt to match
that of GitHub's Copilot Chat. The result is that responses from the
LLMs are as terse and as functional as Copilot Chat's, resulting in less
token usage. I've noticed that with Anthropic as the LLM, the responses
are actually far more impressive. It was able to identify mistakes in
codes that Copilot Chat would only spot when prompted to look for them.

I've also added in new default prompts (and slash commands) for parity
with Copilot Chat which should cover the majority of use cases for
users. As per the README, these can be triggered via the cmd line with
`:CodeCompanion /<slash_command>`.