feat!: use prompt library in chat buffer and better keymap support
Hopefully this release dramatically improves the usability of the
plugin.

Firstly, you can now mark specific prompts (from the prompt library) as
available to appear in the slash command completion menu in the chat
buffer via the `opts.is_slash_cmd` value. This will close #286.

Secondly, the default keymaps that came with each prompt have been
removed in place of `require("codecompanion").prompt("commit")`. This
makes it much cleaner to assign your own keymaps to prompts in the
prompt library.