feat(completion): Boost score of blink completion suggestions
By default, blink's `path` provider has a score offset of 3, meaning
that completions for slash commands in the buffer are deep down
below possible path completions in the list of suggestions.

Since this suggestion only triggers in the chat buffer, it is reasonable
to assume that slashcommands are much more likely to be the correct
suggestions than folders at `/`.