fix: delete diff buffers instead of closing
When closed, a diff buffer still exists (it's just hidden and unlisted).
It also keeps "diffthis" set. This can lead to complicated and confusing
diffs as the number of dangling buffers increases.

Since diff buffers are temporary and not re-used it should be safe to
delete them instead of closing them.