perf(copilot): cache models available for copilot adapter
The first time the copilot adapter is used, it will perform the request
to `/models` endpoint to get the list of available models. This list is
cached and used when the copilot adapter is used again (e.g. when
opening a new chat buffer). To update the list of available models, the
user needs to restart neovim or reload codecompanion plugin.