feat: allow tool_output messages to be rendered in the chat buffer
* When we save the chat.messages and try to pass these on Chat:new()
which chat.ui:render() them, there is no indication that this is a tool
output message.
* This adds the "Tool Output" heading similar how add_tool_message()
appends the content to the buffer.

Fix: for llm tool_calls msg.content will be nil. When we try to pass
messages that have tool_calls field in the message but no content, this
throws an error hence `msg.content or ""`