Allowed flexibility of spaces in the diffs
The diffs format is not standardised.

Python recommends spaces after `+`/`-` in diffs:
https://docs.python.org/3/library/difflib.html#difflib.Differ

GNU's udiff uses no spaces after `+`/`-`
https://www.gnu.org/software/diffutils/manual/html_node/Detailed-Unified.html

It is better to be flexible and allow both approaches rather than trying
to get a standard output using prompts.

Somehow, Anthropic always added spaces. While GPT models never added
spaces. This seems to be because of the difference in the datasets they
were trained on.