feat(chat): add_buffer
This is feature to expose `add_buffer` as public API. E.g. it can be
used to add current buffer to the most recent chat like this:

```lua
function()
  local bufnr = vim.api.nvim_get_current_buf()
  require('codecompanion').last_chat():add_buffer({ bufnr = bufnr })
end
```

Change itself is quite simple. I simply moved code from
`slash_commands/buffer.lua` to `init.lua`, added path resolution in case
it is not supplied and documented it in comments. As well moved code is
called in `slash_commands/buffers.lua` output. Path resolution was
added to make `add_buffer` usage simpler.

This is part of functionality discussed here:
https://github.com/olimorris/codecompanion.nvim/discussions/1546