fix(symbols): also use file contents to detect filetype
This makes the `/symbols` slash command work with files that don't have
an extension, e.g. a Python script without `.py`.

Since we already load the file contents to get the symbols we can just
pass it to `vim.filetype.match` to improve the detection.

Use `unpack()` in the tests to get multiple return values through
`lua_get()`, which only seems to support returning a single value.