Fix inline adapter parsing after ACP integration
The change from adapters to adapters.http and adapters.acp broke the inline parsing for adapters when you do
```
:CodeCompanion <adapter> prompt
```

I used the same "hack" that is used inside adapters/init.lua to generate the full table of possible adapters. I also added a comment to change it as soon as v18 ships. 

I do not have time to look at adding a test for this, I only checked manually with both default and custom adapters and it works.